Query observation data

Function queries the CDEC site to obtain desired station data based on station, sensor number, duration code and start/end date. Use cdec_datasets() to view an updated list of all available data at a station.

cdec_query(station, sensor_num = NULL, dur_code = NULL, start_date = NULL,
  end_date = NULL)

Arguments

station
three letter identification for CDEC location (example "KWK", "SAC", "CCR")

sensor_num
sensor number for the measure of interest. (example "20", "01", "25")

dur_code
duration code for measure interval, "E", "H", "D", which correspong to Event, Hourly and Daily.

start_date
date to start the query on.

end_date
a date to end query on, defaults to current date.

Value

dataframe

Examples

kwk_hourly_flows <- CDECRetrieve::cdec_query("KWK", "20", "H", "2017-01-01")

#> Parsed with column specification:
#> cols(
#>   X1 = col_character(),
#>   X2 = col_character(),
#>   X3 = col_integer(),
#>   X4 = col_character(),
#>   X5 = col_character(),
#>   X6 = col_character(),
#>   X7 = col_character(),
#>   X8 = col_character()
#> )

ccr_hourly_temps <- CDECRetrieve::cdec_query("CCR", "25", "H", Sys.Date())

#> Parsed with column specification:
#> cols(
#>   X1 = col_character(),
#>   X2 = col_character(),
#>   X3 = col_integer(),
#>   X4 = col_character(),
#>   X5 = col_character(),
#>   X6 = col_character(),
#>   X7 = col_character(),
#>   X8 = col_character()
#> )