List Rating Tables

Get a list of all rating tables available through CDEC

cdec_rt_list(station_id = NULL)

Arguments

station_id
station for the location to get rating description for.

Examples

# list all rating tables in CDEC, you can use filter to search
cdec_rt_list()

#> # A tibble: 305 x 4
#>    station_name                         station_id last_revised table_type
#>    <chr>                                <chr>      <date>       <chr>     
#>  1 a&b canal nr johnstonville           abj        2005-07-10   loglog2flo
#>  2 alameda ck above div dam nr sunol ca acd        2019-03-20   usgsrate  
#>  3 alameda ck nr niles                  aln        2019-03-20   usgsrate  
#>  4 american river at chili bar          cbr        2019-03-01   usgsrate  
#>  5 american river at fair oaks          afo        2019-03-20   usgsrate  
#>  6 arcade ck nr del paso hts            ack        2019-03-20   usgsrate  
#>  7 arroyo seco nr soledad               ass        2019-03-20   usgsrate  
#>  8 arroyo valle nr livermore            avl        2019-03-20   usgsrate  
#>  9 battle creek                         bat        2019-03-20   usgsrate  
#> 10 bear ck ab holsten cyn nr rumsey     brk        2019-03-20   usgsrate  
#> # … with 295 more rows