Search CDEC Stations

search the stations in thec CDEC system using the CDEC Station Search service here. Combinations of these parameters can be supplied to refine or be left out to generalize, at least one must be supplied.

cdec_stations(station_id = NULL, nearby_city = NULL, river_basin = NULL,
  hydro_area = NULL, county = NULL)

Arguments

station_id
string three letter station code

nearby_city
string search stations near supplied city

river_basin
string search stations in supplied basin

hydro_area
string search stations in supplied hydrological area

county
string search stations in supplied county

Examples

# cdec_stations() can be used to find locations within an area of interest
cdec_stations(county = "alameda")

#> # A tibble: 27 x 9
#>    station_id name    river_basin county longitude latitude elevation operator 
#>    <chr>      <chr>   <chr>       <chr>      <dbl>    <dbl> <chr>     <chr>    
#>  1 dco        del va… east bay    alame…     -122.     37.6 542 &nbsp CA Dept …
#>  2 oso        oaklan… east bay    alame…     -122.     37.8 1000 &nb… Oakland …
#>  3 chb        chabot  alameda cr  alame…     -122.     37.7 250 &nbsp East Bay…
#>  4 bth        bethan… delta       alame…     -122.     37.8 250 &nbsp CA Dept …
#>  5 cvc        castro… east bay    alame…     -122.     37.7 100 &nbsp US Geolo…
#>  6 aln        alamed… alameda cr  alame…     -122.     37.6 86 &nbsp  US Geolo…
#>  7 cad        calave… alameda cr  alame…     -122.     37.6 1140 &nb… San Fran…
#>  8 usl        u san … alameda cr  alame…     -122.     37.8 477 &nbsp East Bay…
#>  9 ono        oaklan… east bay    alame…     -122.     37.9 1500 &nb… Oakland …
#> 10 nbv        vallec… sf bay      alame…     -122.     37.6 635 &nbsp CA Dept …
#> # ... with 17 more rows, and 1 more variable: state <chr>

# or it can be used to get metadata attributes for a location
cdec_stations(station_id = "ccr")

#> # A tibble: 1 x 9
#>   station_id name     river_basin county longitude latitude elevation operator 
#>   <chr>      <chr>    <chr>       <chr>      <dbl>    <dbl> <chr>     <chr>    
#> 1 ccr        sacrame… sacramento… shasta     -122.     40.5 450 &nbsp US Burea…
#> # ... with 1 more variable: state <chr>